Changed Token to string in nreader and deleted token.go
Design decision change. It was a dark and stormy night. As
lightning struck for the third time we realised that it would
be better to use strings instead of Tokens. As a project intended
for natural language it would not be a limiting factor. Also,
using strings would be more efficient as we no longer have to
do type reflection. It would also integrate better with the
rest of the project. The storm ended. The end.